+++ fiddles/angular2/fiddle-0024-DashboardLayout/src/component/Content.ts
@@ -2,6 +2,8 @@ import {Component, View} from 'angular2/core';
 import * as core from 'angular2/core';
 import {Http, Headers, HTTP_PROVIDERS} from 'angular2/http';
 import {Chart} from './Chart';
+import {TreeView} from './TreeView';
+import {TreeNode} from './TreeNode';
 import {DataJsonApi} from './DataJsonApi';
 
 
@@ -9,56 +11,65 @@ import {DataJsonApi} from './DataJsonApi';
     selector: 'content',
     template: `
      <div class="wrapper" style="height: {{height}}px;">
-        <div class="west">
-            <ng-content select="[west]"></ng-content>
+        <div class="west col-sm-1">
+            <p>Search: <input style="width: 60%;" (change)="onSearchChange()">&nbsp;
+            <a href="#" class="btn btn-primary btn-xs" (click)="onShowSelectedClick">
+                <span class="fa fa-list-ul" title="Show selected"></span>
+            </a>&nbsp;
+            <a href="#" class="btn btn-default btn-xs" (click)="onResetClick">
+                <span class="fa fa-times" title="Unselect all"></span>
+            </a></p>
+            <treeview [store]="treeNodes" [height]="treeViewHeight"></treeview>
         </div>
         <div class="main">
                 <chart *ngFor="#chartObj of charts" [model]='chartObj'
                     style="width: 100%; height: {{mainItemHeight}}px;">
                 </chart>
-                <ag-grid-ng2 class="ag-dark"
-                    style="width: 100%; height: {{mainItemHeight}}px;"
+                <ag-grid-ng2 class="ag-fresh"
+                    style="width: 100%; height: {{mainItemHeight}}px; color: black;"
                     [gridOptions]="gridOptions"
                     [columnDefs]="columnDefs"
                     [rowData] = "rowData">
                 </ag-grid-ng2>
         </div>
     </div>`,
-    directives: [(<any>window).ag.grid.AgGridNg2, Chart],
+    directives: [(<any>window).ag.grid.AgGridNg2, Chart, TreeView],
     providers: [HTTP_PROVIDERS, DataJsonApi]
 })
 export class Content implements OnInit {
     charts:Chart[];
 
     constructor(private dataJsonApi:DataJsonApi) {
-        window.onresize = this.onWindowResize;
-    }
-
-    onWindowResize() {
-        console.log('onWindowResize');
-        var me = this;
-        window.setTimeout(function () {
-
-            $('chart').css('height', me.mainItemHeight);
-            $('#chart-1').css('height', me.mainItemHeight);
-            $('#chart-1-top').css('height', me.mainItemHeight);
-            $('#chart-1-svg').height = me.mainItemHeight;
-            $('#chart-1-svg').css('height', me.mainItemHeight);
-            $('#chart-1-img').css('height', me.mainItemHeight);
-        }, 1000);
+        this.treeNodes = [];
     }
 
     ngOnInit() {
         this.dataJsonApi.request().subscribe((res:Response) => {
-            this.total = res.total;
-            if (res.data) {
-                this.columnDefs = this.parseColumns(res.data[0]);
-                this.configureChart(res.data);
-                this.dataJsonApi.request().subscribe((res:Response) => {
-                    if (res) {
-                        this.rowData = res.data;
+            var accounts = [];
+
+            if (res.children) {
+
+                res.children.map(function(parentNode) {
+                    var childNodes = [];
+                    if(!parentNode.leaf) {
+                        parentNode.children.map(function(child) {
+                            childNodes.push(new TreeNode(child.title, []);
+                        });
                     }
-                });
+                    childNodes.sort(function(a, b) {
+                        return a.name < b.name;
+                    });
+                    this.treeNodes.push(new TreeNode(parentNode.title, childNodes));
+                    if(parentNode.children) {
+                        accounts = accounts.concat(parentNode.children);
+                    }
+                }, this);
+
+
+                this.columnDefs = this.parseColumns(accounts[0]);
+                this.configureChart(accounts);
+                this.rowData = accounts;
+
             }
         });
     }
@@ -85,11 +96,11 @@ export class Content implements OnInit {
                 "scale-x": {
                     "label": {
                         "text": "Checking vs. Savings",
-                        "color": "#ffffff"
+                        "color": "#000000"
 
                     }
                 },
-                backgroundColor: "#333333",
+                backgroundColor: "#ffffcc",
                 tooltip: {
                     padding: 5,
                     offsetY: -5,
@@ -111,9 +122,11 @@ export class Content implements OnInit {
         var columns = [];
         if (rec) {
             Object.keys(rec).map(function (key) {
-                columns.push(
-                    {headerName: key, field: key}
-                );
+                if (key !== 'leaf' && key !== 'title') {
+                    columns.push(
+                        {headerName: key, field: key}
+                    );
+                }
             });
         }
         return columns;
@@ -132,9 +145,23 @@ export class Content implements OnInit {
         return window.innerHeight - 60;
     }
 
+    get treeViewHeight() {
+        return this.height - 50;
+    }
+
     get mainItemHeight() {
         return Math.floor(this.height / 2);
     }
 
+    onSearchChange() {
+        console.log('onSearchChange');
+    }
+    onResetClick() {
+        console.log('onResetClick');
+
+    }
+    onShowSelectedClick() {
+        console.log('onShowSelectedClick');
+    }
 
 }