+++ fiddles/angular2/fiddle-0025-TreeComponent/src/component/Tree.js
@@ -1,0 +1,1 @@
+//# sourceMappingURL=Tree.js.map