+++ fiddles/angular2/fiddle-0025-TreeComponent/src/component/TreeDataService.ts
@@ -1,0 +1,17 @@
+import {Component, enableProdMode, Injectable} from 'angular2/core';
+import {Http, Headers, HTTP_PROVIDERS} from 'angular2/http';
+import 'rxjs/add/operator/map';
+
+
+
+@Injectable()
+export class TreeDataService {
+    constructor(private http: Http) {}
+    request() {
+        const endpoint = metadata.urls.data;
+        return this.http
+            .get(endpoint)
+            .map(res => res.json());
+    }
+}
+