+++ fiddles/angular2/fiddle-0025-TreeComponent/src/component/TreeNodeCheckedPipe.js
@@ -1,0 +1,1 @@
+//# sourceMappingURL=TreeNodeSelectedPipe.js.map