+++ fiddles/angular2/fiddle-0025-TreeComponent/src/component/TreeNodeSearchPipe.js
@@ -1,0 +1,1 @@
+//# sourceMappingURL=TreeViewSearchPipe.js.map