+++ fiddles/angular2/fiddle-0025-TreeComponent/src/meta.js
@@ -1,7 +1,8 @@
 var metadata = {
     fiddleHeader: 'Angular 2 - TreeView',
     urls: {
-        github: 'https://github.com/bradyhouse/house/tree/master/fiddles/angular2/fiddle-0025-TreeView'
+        github: 'https://github.com/bradyhouse/house/tree/master/fiddles/angular2/fiddle-0025-TreeView',
+        data: "data.json"
     },
     consoleTag: 'H O U S E ~ f i d d l e s'
 };